finish ex4.13

+++ part4/blog/controller/blogs.js
@@ -1,7 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../model/blog')
-const User = require('../model/user')
-const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
@@ -12,13 +10,10 @@ blogsRouter.get('/', async (request, response) => {
   
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-    
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id){
-      return response.status(401).json({ error: 'token missing or invalid'})
-    }
 
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
+
+    if(!user) return response.status(401).json({ error: "Unauthorized"})
 
     if (!body.title || !body.url) return response.status(400).end()
 
@@ -39,6 +34,14 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  const user = request.user
+
+  if (blogToDelete.user.toString() !== user._id.toString()){
+    return response.status(401).json({ error: 'token invalid'})
+  }
+  
   await Blog.findByIdAndDelete(request.params.id)
 
   response.status(204).end();
